Conflicts:
	signal-filteroperation.cpp

+++ tfr-cwt.h
@@ -34,7 +34,7 @@ public:
 
     float     min_hz() const { return _min_hz; }
     void      min_hz(float f);
-    float     max_hz(unsigned sample_rate) const { return sample_rate/2; }
+    float     max_hz(unsigned sample_rate) const { return sample_rate/2.f; }
     float     number_of_octaves( unsigned sample_rate ) const;
     unsigned  nScales(unsigned FS) { return (unsigned)(number_of_octaves(FS) * scales_per_octave()); }
     unsigned  scales_per_octave() const { return _scales_per_octave; }
@@ -59,7 +59,7 @@ private:
     Fft             _fft;
     cudaStream_t    _stream;
     float           _min_hz;
-    unsigned        _scales_per_octave;
+    float           _scales_per_octave;
 
     /**
       Default value: wavelet_std_samples=0.03.
